refactor(components): [transfer] use JSX in Unit test
